增GlobalUsing;改size_t

+++ RimeApi/src/RimeStringSlice.cs
@@ -5,5 +5,5 @@ namespace Rime.Api;
 [StructLayout(LayoutKind.Sequential)]
 unsafe public struct RimeStringSlice{
 	public byte* str;// const char
-	public u64 length;
+	public size_t length;
 }
